backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
alpaca-py==0.28.1
sqlalchemy==2.0.35
greenlet>=3,<4
aiosqlite==0.20.0
pandas>=1.5,<3
numpy>=1.24,<3
ta==0.11.0
websockets==13.0.1
python-dotenv==1.0.1
